+++ celescope/fl_vdj_CR/__init__.py
@@ -1,7 +1,0 @@
-__STEPS__ = ['sample', 'barcode', 'convert', 'assemble', 'annotation', 'match', 'summarize', 'mapping'] 
-__ASSAY__ = 'fl_vdj_CR'
-
-
-
-
-