+++ celescope/flv_CR/__init__.py
@@ -1,0 +1,8 @@
+__STEPS__ = ['sample', 'barcode', 'convert', 'assemble', 'summarize', 'match', 'mapping'] 
+__ASSAY__ = 'flv_CR'
+
+
+
+
+
+